fedora 27: almost completely working
this is a changeover from using docker images to use yum/chroot.
obvious issues with this approach - but unlike the docker version
where the issue was loading init, right now the issue is getting
yum to install in the chroot.

will finish up tomorrow.

note that this limits the deployments to debian, have not tested
on rhel systems